Lesson 4.1 Assignment fixed
